Refactor ChartViews
